Added URI's for candidates were possible.
Of the 728 candidates there are DB URI's for 66.

These, plus the hundred or so incorrect matches which were weeded out, have all been reviewed manually and correlated to check they refer to that MP (and not to an MP of another party with the same name; living or dead).

The JSON list has also been cleaned up slightly.